+++ injectool/core.py
@@ -11,10 +11,13 @@ class DependencyError(Exception):
 
 
 def get_dependency_key(dependency: Union[str, Callable]) -> str:
+    """returns string key for passed dependency"""
     return dependency if isinstance(dependency, str) else dependency.__name__
 
 
 class Resolver:
+    """Interface for resolver"""
+
     @abstractmethod
     def resolve(self, container: 'Container', param: Any = None):
         """Factory method for resolving dependency"""
@@ -26,11 +29,10 @@ class SingletonResolver(Resolver):
     def __init__(self, value: Any = None, param: Any = None):
         self._values = {}
         if value is not None:
-            self.add_value(value, param)
+            self.set_value(value, param)
 
-    def add_value(self, value: Any, param: Any):
-        if param in self._values:
-            raise DependencyError(f'Singleton value for parameter {param} is already added')
+    def set_value(self, value: Any, param: Any):
+        """Sets value for parameter"""
         self._values[param] = value
 
     def resolve(self, container: 'Container', param: Any = None):
@@ -65,6 +67,7 @@ class Container:
 
     @property
     def name(self) -> str:
+        """Returns container's name"""
         return self._name
 
     def add(self, dependency: Union[str, Callable], resolver: Resolver):
@@ -82,6 +85,7 @@ class Container:
             raise DependencyError('Dependency "{0}" is not found'.format(key))
 
     def get_resolver(self, dependency: Union[str, Callable]) -> Resolver:
+        """returns resolver for dependency"""
         key = get_dependency_key(dependency)
         return self._resolvers.get(key, None)
 
@@ -94,6 +98,7 @@ class Container:
 
     @staticmethod
     def get(name: str = None) -> 'Container':
+        """Returns container by name or default if name is None"""
         try:
             return Container._get_default() if name is None else Container._containers[name]
         except KeyError:
@@ -113,6 +118,10 @@ Container('')
 
 @contextmanager
 def make_default(container_name: str) -> Container:
+    """
+    Makes container with provided name default.
+    Creates new if container doesn't exist.
+    """
     try:
         container = Container.get(container_name)
     except DependencyError:
@@ -135,14 +144,15 @@ def add_singleton(dependency: Union[str, Callable], value: Any):
     Container.get().add(dependency, SingletonResolver(value))
 
 
-def add_resolve_function(dependency: Union[str, Callable], resolve_: Callable[['Container', Any], Any]):
+def add_function_resolver(dependency: Union[str, Callable],
+                          resolve_: Callable[['Container', Any], Any]):
     """Adds function resolver to current container"""
     Container.get().add(dependency, FunctionResolver(resolve_))
 
 
 def add_type(dependency: Union[str, Callable], type_: Type):
     """Adds type resolver to current container"""
-    add_resolve_function(dependency, lambda c, p=None: type_())
+    add_function_resolver(dependency, lambda c, p=None: type_())
 
 
 def resolve(dependency: Union[str, Callable], param: Any = None):