+++ injectool/injection.py
@@ -3,11 +3,14 @@
 from functools import wraps
 from typing import List, Union, Callable
 
-from injectool.core import get_dependency_key, resolve
+from injectool.core import get_dependency_key, resolve, DependencyError
 
 
 def inject(*dependencies: List[Union[str, Callable]]):
-    """Resolves dependencies in default container and passes it as optional parameters to function"""
+    """
+    Resolves dependencies in default container
+    and passes it as optional parameters to function
+    """
 
     def _decorate(func):
         keys = [get_dependency_key(dep) for dep in dependencies]
@@ -25,16 +28,15 @@ def inject(*dependencies: List[Union[str, Callable]]):
     return _decorate
 
 
-def dependency(dep: Union[str, Callable] = None):
+def dependency(func):
     """Substitute function by calling resolved in default container"""
 
-    def _decorate(func):
-        key = func if dep is None else dep
+    @wraps(func)
+    def _decorated(*args, **kwargs):
+        try:
+            implementation = resolve(func)
+        except DependencyError:
+            implementation = func
+        return implementation(*args, **kwargs)
 
-        @wraps(func)
-        def _decorated(*args, **kwargs):
-            return resolve(key)(*args, **kwargs)
-
-        return _decorated
-
-    return _decorate
+    return _decorated