+++ injectool/tests/core_tests.py
@@ -4,7 +4,7 @@ from pytest import raises, mark, fixture, fail
 
 from injectool import Container, DependencyError
 from injectool.core import get_dependency_key, SingletonResolver, FunctionResolver, add_type
-from injectool.core import make_default, add_resolver, add_singleton, add_resolve_function, resolve
+from injectool.core import make_default, add_resolver, add_singleton, add_function_resolver, resolve
 
 
 @mark.parametrize('dependency, key', [
@@ -30,23 +30,10 @@ class SingletonResolverTests:
         """should add single value for parameter"""
         resolver = SingletonResolver(def_value, def_param)
 
-        resolver.add_value(value, param)
+        resolver.set_value(value, param)
 
         assert resolver.resolve(Mock(), def_param) == def_value
         assert resolver.resolve(Mock(), param) == value
-
-    @staticmethod
-    @mark.parametrize('value, param', [
-        (2, SingletonResolver),
-        ({'key': 1}, None),
-        (['value'], 'parameter')
-    ])
-    def tests_add_raises_for_exiting_parameter(value, param):
-        """should raise error for adding existing param"""
-        resolver = SingletonResolver(value, param)
-
-        with raises(DependencyError):
-            resolver.add_value(value, param)
 
     @staticmethod
     @mark.parametrize('param', [SingletonResolver, None, 'parameter'])
@@ -78,46 +65,6 @@ class FunctionResolverTests:
         resolver = FunctionResolver(lambda c, p=None: value)
 
         assert resolver.resolve(container) == value
-
-
-class TypeResolverTests:
-    """Type resolver class tests"""
-
-    @staticmethod
-    @mark.parametrize('def_value, def_param, value, param', [
-        (1, None, 2, SingletonResolver),
-        ({}, SingletonResolver, {'key': 1}, None),
-        ([], 1, ['value'], 'parameter')
-    ])
-    def tests_adds_type_for_parameter(def_value, def_param, value, param):
-        """should add single value for parameter"""
-        resolver = SingletonResolver(def_value, def_param)
-
-        resolver.add_value(value, param)
-
-        assert resolver.resolve(Mock(), def_param) == def_value
-        assert resolver.resolve(Mock(), param) == value
-
-    @staticmethod
-    @mark.parametrize('value, param', [
-        (2, SingletonResolver),
-        ({'key': 1}, None),
-        (['value'], 'parameter')
-    ])
-    def tests_add_raises_for_exiting_parameter(value, param):
-        """should raise error for adding existing param"""
-        resolver = SingletonResolver(value, param)
-
-        with raises(DependencyError):
-            resolver.add_value(value, param)
-
-    @staticmethod
-    @mark.parametrize('param', [SingletonResolver, None, 'parameter'])
-    def test_resolve_raises_for_unknown_param(param):
-        resolver = SingletonResolver()
-
-        with raises(DependencyError):
-            resolver.resolve(param)
 
 
 @fixture
@@ -358,10 +305,10 @@ def test_add_singleton(dependency, value):
     ('key', lambda c, p=None: 'value'),
     (get_dependency_key, lambda c, p=None: 1)
 ])
-def test_add_resolve_function(dependency, function):
-    """add_function() should add FunctionResolver to current container"""
+def test_add_function_resolver(dependency, function):
+    """should add FunctionResolver to current container"""
     with make_default('test_add_resolve_function') as container:
-        add_resolve_function(dependency, function)
+        add_function_resolver(dependency, function)
 
         resolver = container.get_resolver(dependency)
         assert isinstance(resolver, FunctionResolver)