+++ injectool/tests/injection_tests.py
@@ -2,18 +2,18 @@ from unittest.mock import Mock
 
 from pytest import mark, fixture
 
-from injectool.core import make_default, Container, get_dependency_key, add_singleton, add_resolve_function
+from injectool.core import make_default, Container, get_dependency_key, add_singleton
 from injectool.injection import inject, dependency
 
 
 @fixture
 def inject_fixture():
-    name = 'test_container'
-    Container._containers.pop(name, None)
+    name = 'inject_tests'
     with make_default(name) as container:
         yield container
 
 
+@mark.usefixtures('inject_fixture')
 @mark.parametrize('dep, key', [
     ('key', 'key'),
     (get_dependency_key, 'get_dependency_key'),
@@ -32,34 +32,21 @@ def test_inject(dep, key):
     assert get_implementation() == value
 
 
-@dependency()
-def interface_func():
-    pass
-
-
-@dependency('func_key')
-def keyed_func():
-    pass
-
-
-class Interface:
-    pass
-
-
-@dependency(Interface)
-def get_interface():
-    pass
-
-
-@mark.parametrize('dep, key', [
-    (interface_func, get_dependency_key(interface_func)),
-    (keyed_func, 'func_key'),
-    (get_interface, get_dependency_key(Interface)),
-])
-def test_dependency(dep, key):
+@mark.usefixtures('inject_fixture')
+@mark.parametrize('default', [True, False])
+def test_dependency(default):
+    """should resolve function or use decorated in case not registered"""
     implementation = Mock()
-    add_resolve_function(key, lambda c, param=None: implementation)
+    default_implementation = Mock()
 
-    dep()
+    @dependency
+    def func():
+        default_implementation()
 
-    assert implementation.called
+    if not default:
+        add_singleton(func, implementation)
+
+    func()
+
+    assert default_implementation.called == default
+    assert implementation.called == (not default)