+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Dev
+
+- @dependency uses own implementation by default
+- add_resolve_function is renamed to add_function_resolver
 
 ## 1.1.1